Pick an engine-specific value

Set different values that will be used based on the current epoxy or knitr engine (one of md, html, or latex). The engine-specific value will be used inside epoxy knitr chunks or epoxy functions matching the source syntax: epoxy() (md), epoxy_html() (html), or epoxy_latex() (latex).

Usage

engine_pick(md, html = md, latex = md)

Arguments

md, html, latex

The value to use in a markdown, HTML, or LaTeX context.

Value

The value of md, html or latex depending on the epoxy or knitr currently being evaluated.

Examples

# Markdown and HTML are okay with bare `$` character,
# but we need to escape it in LaTeX.
engine_pick(md = "$", latex = "\\$")
#> [1] "$"