Style replacements with inline formatting

Makes available all of the label_ functions from the scales package within an inline format function, fmt(). To apply formatting, wrap embraced expressions in fmt(expr, label), where label is the name of a labeller function from scales (with or without the leading label_), e.g. "dollar" or "label_dollar" for scales::label_dollar(). Short label forms exist for some labellers, e.g. "$" for "dollar". See the full listing in the table below.

```{r}
# Airbnb stats that may not be completely accurate
airbnb <- list(
 countries = 220,
 cities = 1e5,
 avg_stay = 4.326,
 avg_cost = 184.952,
 hosts = 4.12e6
)

With standard inline R code in R Markdown, we can write the following:

* Airbnb includes listings in `r airbnb$cities` cities
  in `r airbnb$countries` countries
  from around `r airbnb$hosts` hosts.
  The average guest stays `r airbnb$avg_stay` nights
  at a rate of `r airbnb$avg_cost` per night.

Airbnb includes listings in 10^5 cities in 220 countries from around 4.12 × 10^6 hosts. The average guest stays 4.326 nights at a rate of 184.952 per night.

Using epoxy and the light-weight fmt() function from epoxy_style_format(), we can improve the readability and formatting of the interwoven numbers.

```{r my_style, echo = FALSE}
# Define number/dollar styles to apply consistently
my_style <- epoxy_style_format(
 dollar = scales::label_dollar(accuracy = 1),
 number = scales::label_number(
   accuracy = 0.1,
   scale_cut = scales::cut_short_scale()
 )
)

Airbnb includes listings in 100,000 cities in 220 countries from around 4.1M hosts. The average guest stays 4.3 nights at a rate of $185 per night.

fmt() labels for formatters

label	Applies formatting with

"bytes"	scales::label_bytes()

"d", "date"	scales::label_date()

"ds", "date_short"	scales::label_date_short()

"dt", "time"	scales::label_time()

"$", "dollar"	scales::label_dollar()

"log"	scales::label_log()

"#", "number"	scales::label_number()

",", "comma"	scales::label_comma()

"a", "auto", "number_auto"	scales::label_number_auto()

"o", "ordinal"	scales::label_ordinal()

"parse"	scales::label_parse()

"math"	scales::label_math()

"pct", "%", "percent"	scales::label_percent()

"p", "pvalue"	scales::label_pvalue()

"si", "scientific"	scales::label_scientific()

"wrap"	scales::label_wrap()

"uc", "uppercase"	toupper()

"lc", "lowercase"	tolower()

"tc", "titlecase"	tools::toTitleCase()

Usage

epoxy_style_format(
  bytes = scales::label_bytes(),
  date = scales::label_date(),
  date_short = scales::label_date_short(),
  time = scales::label_time(),
  dollar = scales::label_dollar(),
  log = scales::label_log(),
  number = scales::label_number(),
  comma = scales::label_comma(),
  number_auto = scales::label_number_auto(),
  ordinal = scales::label_ordinal(),
  parse = scales::label_parse(),
  math = scales::label_math(),
  percent = scales::label_percent(),
  pvalue = scales::label_pvalue(),
  scientific = scales::label_scientific(),
  wrap = scales::label_wrap(width = 80),
  uppercase = toupper,
  lowercase = tolower,
  titlecase = tools::toTitleCase,
  ...,
  transformer = glue::identity_transformer
)

Arguments

bytes

The function to apply to when label is "bytes". Default is scales::label_bytes().

date

The function to apply to when label is "date" or "d". Default is scales::label_date().

date_short

The function to apply to when label is "date_short" or "ds". Default is scales::label_date_short().

time

The function to apply to when label is "time" or "dt". Default is scales::label_time().

dollar

The function to apply to when label is "dollar" or "$". Default is scales::label_dollar().

log

The function to apply to when label is "log". Default is scales::label_log().

number

The function to apply to when label is "number" or "#". Default is scales::label_number().

comma

The function to apply to when label is "comma" or ",". Default is scales::label_comma().

number_auto

The function to apply to when label is "number_auto", "a", or "auto". Default is scales::label_number_auto().

ordinal

The function to apply to when label is "ordinal" or "o". Default is scales::label_ordinal().

parse

The function to apply to when label is "parse". Default is scales::label_parse().

math

The function to apply to when label is "math". Default is scales::label_math().

percent

The function to apply to when label is "percent", "pct", or "%". Default is scales::label_percent().

pvalue

The function to apply to when label is "pvalue" or "p". Default is scales::label_pvalue().

scientific

The function to apply to when label is "scientific" or "si". Default is scales::label_scientific().

wrap

The function to apply to when label is "wrap". Default is scales::label_wrap().

uppercase

The function to apply to when label is "uppercase" or "uc". Default is toupper().

lowercase

The function to apply to when label is "lowercase" or "lc". Default is tolower().

titlecase

The function to apply to when label is "titlecase" or "tc". Default is tools::toTitleCase().

...

Additional formatting functions as named arguments. The name of the argument in ... determines the label value associated with the formatter in fmt().

For example, providing url = utils::URLencode would allow you to apply URL-encoding formatting using fmt(expr, "url").

transformer

The transformer to apply to the replacement string. This argument is used for chaining the transformer functions. By providing a function to this argument you can apply an additional transformation after the current transformation. In nearly all cases, you can let epoxy_style() handle this for you. The chain ends when glue::identity_transformer() is used as the transformer.

Examples

glue::glue(
  '{fmt(0.2, "%")} of revenue generates {fmt(42000, "$")} in profits.',
  .transformer = epoxy_style_format()
)
#> 20% of revenue generates $42,000 in profits.