Epoxy Inline Style Transformer

This epoxy style is heavily inspired by the inline formatters in the cli package. The syntax is quite similar, but epoxy's syntax is slightly different to accomodate reporting use cases.

With the inline styles, you can include a keyword, prefixed with a dot (.) that is used to format the template variable in place.

epoxy("It cost {.dollar 123456}.", .style = "inline")
#> It cost $123,456.

The formatters, e.g. dollar in the example above, can be customized using the arguments of epoxy_style_inline(). Pass a customized scales::label_dollar() to dollar to achieve a different style.

dollars_nzd <- scales::label_dollar(suffix = " NZD")

epoxy(
  "It cost {.dollar 123456}.",
  .style = epoxy_style_inline(dollar = dollars_nzd)
)
#> It cost $123,456 NZD.

Note that, unlike inline markup with cli, the text within the template variable, other than the keyword, is treated as an R expression.

money <- 123456
epoxy("It cost {.dollar money}.", .style = "inline")
#> It cost $123,456.

You can also nest inline markup expressions.

money <- c(123.456, 234.567)
epoxy("It will cost either {.or {.dollar money}}.", .style = "inline")
#> It will cost either $123.46 or $234.57.

Finally, you can provide your own function that is applied to the evaluated expression.

set.seed(4242)

epoxy(
  "Here are three random percentages: {.and {.pct {.runif 3}}}.",
  .style = epoxy_style_inline(
    runif = function(n) sort(runif(n))
  )
)
#> Here are three random percentages: 23%, 35% and 99%.

Usage

epoxy_style_inline(
  ...,
  transformer = glue::identity_transformer,
  bytes = scales::label_bytes(),
  date = function(x) format(x, format = "%F"),
  time = function(x) format(x, format = "%T"),
  datetime = function(x) format(x, format = "%F %T"),
  dollar = scales::label_dollar(prefix = default_for_engine("$", "$", "\\$")),
  number = scales::label_number(),
  comma = scales::label_comma(),
  ordinal = scales::label_ordinal(),
  percent = scales::label_percent(suffix = default_for_engine("%", "%", "\\%")),
  pvalue = scales::label_pvalue(),
  scientific = scales::label_scientific(),
  uppercase = toupper,
  lowercase = tolower,
  titlecase = tools::toTitleCase
)

Arguments

...

Additional named inline transformers. The evaluated expression from the template expression is passed as the first argument to the function.

transformer

[function]
A function taking three parameters code, envir and data used to transform the output of each block before, during, or after evaluation. For example transformers see vignette("transformers").

bytes

The function to apply to x when the template is {.bytes x}. Default is scales::label_bytes().

date

The function to apply to x when the template is {.date x}. Default is function().

time

The function to apply to x when the template is {.time x}. Default is function().

datetime

The function to apply to x when the template is {.datetime x} or {.dttm x}. Default is function().

dollar

The function to apply to x when the template is {.dollar x}. Default is scales::label_dollar().

number

The function to apply to x when the template is {.number x} or {.num x}. Default is scales::label_number().

comma

The function to apply to x when the template is {.comma x}. Default is scales::label_comma().

ordinal

The function to apply to x when the template is {.ordinal x}. Default is scales::label_ordinal().

percent

The function to apply to x when the template is {.percent x} or {.pct x}. Default is scales::label_percent().

pvalue

The function to apply to x when the template is {.pvalue x}. Default is scales::label_pvalue().

scientific

The function to apply to x when the template is {.scientific x}. Default is scales::label_scientific().

uppercase

The function to apply to x when the template is {.uppercase x} or {.uc x}. Default is toupper().

lowercase

The function to apply to x when the template is {.lowercase x} or {.lo x}. Default is tolower().

titlecase

The function to apply to x when the template is {.titlecase x} or {.tc x}. Default is tools::toTitleCase().

Value

A function of text and envir suitable for the .transformer

argument of glue::glue().

See also